Add -timelimit option
This option limits the CPU time used by ffmpeg to the number of seconds
specified.  After this time, the OS sends a SIGXCPU signal, which we
handle and attempt to exit cleanly.  If the process is stuck, the OS
will deliver a SIGKILL one second later, forcibly terminating the
process.

This functionality is useful in automated setups where a runaway ffmpeg
process would otherwise go undetected.

git-svn-id: svn://svn.mplayerhq.hu/ffmpeg/trunk@21347 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b