DECLARE_ALIGNED macro for TMS470 compiler
This compiler supports gcc-style alignment attributes for struct,
but not for global variables.  For the latter, alignment can be
specified with a pragma, which does not work for struct members or
local variables.  By using both pragma and attribute, one or the
other will always take effect.

Unfortunately, no means exists for aligning stack variables.

git-svn-id: svn://svn.mplayerhq.hu/ffmpeg/trunk@21379 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b