mpegaudio: call ff_mpegaudiodec_init_mmx() only from float decoder
The mmx code is floating-point only, and this function does not know
from which decoder it is called.  Without this change, the integer
decoder only "works" because the size of the context struct is smaller
in this case, and the mmx init function writes the function pointer
outside the allocated context.

git-svn-id: svn://svn.mplayerhq.hu/ffmpeg/trunk@23949 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b