PPC: convert Altivec FFT to pure assembler
On PPC a leaf function has a 288-byte red zone below the stack pointer,
sparing these functions the chore of setting up a full stack frame.

When a function call is disguised within an inline asm block, the
compiler might not adjust the stack pointer as required before a
function call, resulting in the red zone being clobbered.

Moving the entire function to pure asm avoids this problem and also
results in somewhat better code.

git-svn-id: svn://svn.mplayerhq.hu/ffmpeg/trunk@24044 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b