Never shrink the ByteIOContext buffer in ff_rewind_with_probe_data
If there is little unread data in the ByteIOContext buffer, this may lead
to reducing the size of the ByteIOContext buffer to little more the probe
data size. This can lead to suboptimal aviobuf behaviour, e.g. making some
demuxers fail to do short seeks backwards (if the input isn't seekable).


git-svn-id: svn://svn.mplayerhq.hu/ffmpeg/trunk@24478 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b